requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.9
charset-normalizer==3.4.2
diff-match-patch==20241021
Django
django-autoslug==1.9.9
django-cors-headers
django-extensions==4.1
django-import-export==4.3.8
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
idna==3.10
import-export==0.3.1
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications
numpy
openpyxl==3.1.5
panda==0.3.1
pandas==2.3.1
pillow==11.3.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
loguru==0.7.3 ###
pydantic==2.11.7
langchain_community==0.3.27
langchain_core==0.3.74
langchain_openai==0.3.30
langgraph==0.6.6 ####
gunicorn==20.1.0